share metrics in different way
