+++ apps/avatax/src/lib/otel-metrics.ts
@@ -1,7 +1,6 @@
-import { metrics } from "@opentelemetry/api";
-
 import pkg from "../../package.json";
+import { meterProvider } from "./shared-metrics";
 
-export const internalMeter = metrics.getMeter("saleor.app.avatax.core", pkg.version);
+export const internalMeter = meterProvider.getMeter("saleor.app.avatax.core", pkg.version);
 
-export const externalMeter = metrics.getMeter("saleor.app.avatax.service", pkg.version);
+export const externalMeter = meterProvider.getMeter("saleor.app.avatax.service", pkg.version);
