+++ apps/avatax/src/otel-instrumentation.node.ts
@@ -1,35 +1,34 @@
+import { metrics } from "@opentelemetry/api";
 import { W3CTraceContextPropagator } from "@opentelemetry/core";
 import { NodeSDK } from "@opentelemetry/sdk-node";
 import { createBatchSpanProcessor } from "@saleor/apps-otel/src/batch-span-processor-factory";
 import { createHttpInstrumentation } from "@saleor/apps-otel/src/http-instrumentation-factory";
-import { createMetricReader } from "@saleor/apps-otel/src/metric-reader-factory";
-import { createResource } from "@saleor/apps-otel/src/resource-factory";
 
 import { env } from "@/env";
 
-import pkg from "../package.json";
+import { otelResource } from "./lib/otel-resource";
+import { meterProvider } from "./lib/shared-metrics";
 
-const metricReader = createMetricReader({
-  accessToken: env.OTEL_ACCESS_TOKEN!,
-  exportIntervalMillis: 60_000,
-});
+/*
+ * const metricReader = createMetricReader({
+ *   accessToken: env.OTEL_ACCESS_TOKEN!,
+ *   exportIntervalMillis: 60_000,
+ * });
+ */
 
 const sdk = new NodeSDK({
-  resource: createResource({
-    serviceName: env.OTEL_SERVICE_NAME,
-    serviceVersion: pkg.version,
-    serviceEnviroment: env.ENV,
-    serviceCommitSha: env.VERCEL_GIT_COMMIT_SHA,
-  }),
+  resource: otelResource,
   textMapPropagator: new W3CTraceContextPropagator(),
   spanProcessor: createBatchSpanProcessor({
     accessToken: env.OTEL_ACCESS_TOKEN!,
   }),
-  metricReader,
   instrumentations: [createHttpInstrumentation()],
 });
 
 sdk.start();
 
-// hack to make the metric reader available globally
-global.currentMetricReader = metricReader;
+/*
+ * hack to make the metric reader available globally
+ * global.currentMetricReader = metricReader;
+ */
+metrics.setGlobalMeterProvider(meterProvider);
